Rename `src batch remote run` to `src batch remote`.
